test: centralize fixtures parseArgs function
