+++ tests/playwright/fixtures/protocol-hint.tsx
@@ -1,28 +1,8 @@
-import process from "node:process";
 import { Box, render } from "ink";
 import React from "react";
 import type { ImageProtocolHint, ImageProtocolName } from "../../../src";
 import Image, { TerminalInfoProvider } from "../../../src";
-
-function parseArgs(args: string[]) {
-  const result: Record<string, boolean | string> = {};
-
-  for (let i = 0; i < args.length; i++) {
-    const arg = args[i];
-
-    if (arg.startsWith("--")) {
-      const key = arg.slice(2);
-
-      if (i + 1 < args.length && !args[i + 1].startsWith("--")) {
-        result[key] = args[++i];
-      } else {
-        result[key] = true;
-      }
-    }
-  }
-
-  return result;
-}
+import { parseArgs } from "./utils.js";
 
 const parsed = parseArgs(process.argv.slice(2));
 const imagePath = (parsed.src as string) || "";
