+++ tests/playwright/fixtures/utils.ts
@@ -1,0 +1,19 @@
+export function parseArgs(args: string[]) {
+  const result: Record<string, boolean | string> = {};
+
+  for (let i = 0; i < args.length; i++) {
+    const arg = args[i];
+
+    if (arg.startsWith("--")) {
+      const key = arg.slice(2);
+
+      if (i + 1 < args.length && !args[i + 1].startsWith("--")) {
+        result[key] = args[++i];
+      } else {
+        result[key] = true;
+      }
+    }
+  }
+
+  return result;
+}
